+++ src/test/java/video/api/java/sdk/infrastructure/unirest/live/LiveStreamStreamClientTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import video.api.java.sdk.domain.exception.ResponseException;
 import video.api.java.sdk.domain.live.LiveStream;
-import video.api.java.sdk.infrastructure.unirest.asset.AssetsJsonSerializer;
 import video.api.java.sdk.infrastructure.unirest.request.RequestBuilderFactory;
 import video.api.java.sdk.infrastructure.unirest.video.TestRequestExecutor;
 
@@ -17,7 +16,7 @@ class LiveStreamStreamClientTest {
     void setUp() {
         liveStreamClient = new LiveStreamClient(
                 new RequestBuilderFactory(""),
-                new LiveStreamJsonSerializer(new AssetsJsonSerializer()),
+                new LiveStreamJsonSerializer(),
                 new TestRequestExecutor()
         );
     }