+++ CHANGELOG.md
@@ -11,7 +11,7 @@ All notable changes to this project will be documented in this file.
 - `Info.loadedAt` property type from `String` to `Calendar`
 - `Info.endedAt` property type from `String` to `Calendar`
 - `PlayerSessionEvent.emittedAt` property type from `String` to `Calendar`
-- `Client.videos.upload(String source)` -> `Client.videos.upload(File file)`
+- `Client.videos.upload(String sourceInfo)` -> `Client.videos.upload(File file)`
 - `Client.videos.uploadThumbnail(Video video, String thumbnailSource)` -> `Client.videos.uploadThumbnail(Video video, File file)`
 - `Client.videos.uploadThumbnailWithTimecode(Video video, String timecode)` -> `Client.videos.uploadThumbnail(Video video, String timecode)`
 - `Client.liveStreams.uploadThumbnail(String liveStreamId, String thumbnailSource)` -> `Client.liveStreams.uploadThumbnail(String liveStreamId, File file)`